--- src/reducers/filter.js
@@ -1,4 +1,0 @@
-let initialState = {dummy: false};
-export function filter(state = initialState, action) {
-  return state;
-}
