--- src/reducers/index.js
@@ -1,4 +1,0 @@
-let initialState = {dummy: false};
-export let fucks = function (state = initialState, action) {
-  return {spaz: 'bott'};
-}
